LinePrepOpt.patch by Ticker Berkin - improve efficency of bitstream optimization in LinePreparer
"If starting with unsigned deltas in polyline encoding, and attempting
to reduce the length by testing reduced x/yBase values, there isn't any
point in testing Base-1, as the normal number of bits will be the same
(added sign bit) and there will be some overflows.

Patch attached to change this - it gives me a very slight reduction in
RGN size and might give a performance improvement (I didn't detect
any).

I've also fixed a couple of log. logic errors and changed minDx/y
initial to 0, which is much less confusing because they are symmetric
with maxDx/y which do use 0."


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4645 25d90789-57f7-4ee0-8453-03a3dfeeeb22