(modified) mainlogging.patch by Mike Baggaley - changed runAsProcess() to work on unix like systems and without dist - changed wait to 1 MILLISECONDS Original comment by Mike: "Please find attached a patch with the remainder of the logging changes (apart from CommandArgs).
There are changes to the main program so that the dates and times are only output if mkgmap is processing one or more files.

The patch includes an update to the unit test. Note that this now takes a little longer to run, because of having to create new processes to run some of the tests."

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4656 25d90789-57f7-4ee0-8453-03a3dfeeeb22