improve RoadMerger (based on merge-roads-reversed-v6.patch) This has two effects: - it allows RoadMerger to merge roads which build a sharp angle, if that is done an info message is logged - it allows RoadMerger to merge roads with different directions be reversing the order of points if special criteria are met - implement new special tag mkgmap:has-direction=true which tells mkgmap that the direction of the way is important and should not be reversed. The routing tags oneway=yes or oneway=-1 have the same effect, but they also effect routing. Use this tag if you have a TYP file which renders lines different (e.g. cycleway on the right or on the left side) to prevent a reverse merge - implements new experimental option --x-force-reverse-merge
Merging with reverse produces smaller maps, esp. smaller NET and NOD files, it may have an influence on routing.
Reversing of roads is disabled unless mkgmap detects that the style has at least one rule which adds mkgmap:has-direction=true in rule file lines or option --x-force-reverse-merge is used.
The program checks if any line that was added for a given OSM way has the tag mkgmap:has-direction=true.
If one is found the tag is propagated to the road(s) for that OSM way so no reversing happens.
Roads with different direction attributes are not merged.

The option --x-force-reverse-merge is meant for those who don't use a TYP file or don't render left/right attributes with different types. All others should not use the option and adapt the style instead.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4703 25d90789-57f7-4ee0-8453-03a3dfeeeb22