add experimental code to implement new option --improve-overview
For each multipolygon with holes(role=inner) that is visible in the overview map a second version is calculated for the max. resolution of the overview map. The Douglas-Peucker filter is used and small holes which would not appear anyway are removed. The second multipolygon is used to render the overview map levels. 
Expected effects:
- much better overview map (almost no wrong 
- much smaller overview map
- increased run time (depends on style and complexity of the multipolygons), ca. 10% for a map of Norway with the default style

By default the option is switched off, this may change if feedback is positive.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4761 25d90789-57f7-4ee0-8453-03a3dfeeeb22